github_upload_min/requirements_native.txt
streamlit==1.37.1
python-docx==1.1.2
pandas==2.2.2
Pillow==10.4.0
pywebview==5.4
